refactor: move facebook google icons to shared ui folder
