+++ apps/journeys-admin/src/components/SignIn/GoogleButton/GoogleButton.tsx
@@ -1,10 +1,9 @@
 import Button from '@mui/material/Button'
 import { GoogleAuthProvider, getAuth, signInWithPopup } from 'firebase/auth'
 import { useRouter } from 'next/router'
-import { ReactElement } from 'react'
+import React, { ReactElement } from 'react'
 import { useTranslation } from 'react-i18next'
-
-import GoogleIcon from './GoogleIcon'
+import { GoogleIcon } from '../../../../../../libs/shared/ui/src/components/icons/GoogleIcon'
 
 export function GoogleButton(): ReactElement {
   const { t } = useTranslation()
